removed fined biggest errors and instead calculate errors of a single data file
